Refactor page and log layouts and add logviewer tool
- Implement new page layout with 8-byte header containing pageLSN
- Redesign log block layout with boundary-based record storage
- Add logviewer command-line tool for inspecting log records

+++ internal/buffer/manager.go
@@ -43,7 +43,7 @@ func (bm *Manager) Available() int {
 	return bm.numAvailable
 }
 
-func (bm *Manager) FlushAll(txnum int) error {
+func (bm *Manager) FlushAll(txnum int64) error {
 	bm.mu.Lock()
 	defer bm.mu.Unlock()
